+++ frontend/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2017-03-30
+### Added
+- Add application global fatal error message box with APP_FATAL_ERROR action.
+
+### Fixed
+- Fix component recreate issue caused by HOC component.
 
 ## 2017-03-29
 ### Added